boom! working flawlessly! now to raise coverage... ;-)
