+++ .travis.yml
@@ -1,0 +1,5 @@
+language: node_js
+node_js:
+  - 0.10
+env:
+  - FASTER=D03B48854B21F354EE4E8AF097B6BBBDFE3861BE0493443AAE8CDDA2A4296566
