+++ package.json
@@ -35,7 +35,7 @@
   "dependencies": {
     "chalk": "^1.0.0",
     "ignored": "^2.0.1",
-    "listdirs": "^2.0.4",
+    "listdirs": "^3.0.0",
     "ps-tree": "0.0.3",
     "socket.io": "^1.3.5",
     "socket.io-client": "^1.3.5",
